Move more things to domain and infra packages
This is in preparation for adding a REST or HTTP user detail provider.
I'm moving code into the `domain` and `infra` modules. The `domain`
module contains the traits to be implemented in the infra packages.
