+++ src/domain/mod.rs
@@ -1,0 +1,4 @@
+//! The domain module contains code to be dependent upon elsewhere in this project. It
+//! should itself not depend on any modules in this project.
+pub mod events;
+pub mod user;
