+++ src/storage/mod.rs
@@ -1,5 +1,3 @@
 mod choose;
 
 pub use choose::{ChoosingVfs, InnerVfs, SbeMeta};
-pub use unftp_sbe_restrict::RestrictingVfs;
-pub use unftp_sbe_rooter::{RooterVfs, UserWithRoot};
